Make service stopping work again
I don't know when it broke, but the problem was due to
RegisterServiceCtrlHandler not working.  That is,
service-control-handler was never called, even though we had
registered it.  When we switch to RegisterServiceCtrlHandlerEx it
magically started working again.

Also, change dwServiceType to remove SERVICE_INTERACTIVE_PROCESS,
since that doesn't work and we don't need it.